sawtooth/tp/requirements.txt
protobuf==3.17.3
sawtooth-sdk==1.2.5
web3==5.24.0
pymongo==3.12.1
python-dotenv